// src/app/models/clasificados.service.ts
export interface Clasificados{
      id: number;
      no: number;
      fecha: string;
      enviado: string;
      rsb: string;
      rs: string;
      fecha_registro_ctc: string;
      asunto: string;
      doc: string;
      ej: string;
      clasif: string;
      destino: string;
      traslado: string;
      fecha_traslado: string;
      imagen: string;
  }